split `transform_common` into it with `functionalization`
